Move preview panel JS code to a separate file
