+++ wagtail/admin/templates/wagtailadmin/pages/_editor_js.html
@@ -23,6 +23,7 @@
 <script src="{% versioned_static 'wagtailadmin/js/expanding-formset.js' %}"></script>
 <script src="{% versioned_static 'wagtailadmin/js/modal-workflow.js' %}"></script>
 <script src="{% versioned_static 'wagtailadmin/js/page-editor.js' %}"></script>
+<script src="{% versioned_static 'wagtailadmin/js/preview-panel.js' %}"></script>
 <script src="{% versioned_static 'wagtailadmin/js/vendor/xregexp.min.js' %}"></script>
 <script src="{% versioned_static 'wagtailadmin/js/vendor/urlify.js' %}"></script>
 <script src="{% versioned_static 'wagtailadmin/js/workflow-action.js' %}"></script>
